move to double type

+++ Library/DerivativeMethod.cs
@@ -4,9 +4,9 @@ namespace Library
 {
     public class DerivativeMethod
     {
-        public delegate void DerivMethod(float[] prev, float[] newV, float dt, float t, int i, Func<float, float[], float> f_i);
+        public delegate void DerivMethod(double[] prev, double[] newV, double dt, double t, int i, Func<double, double[], double> f_i);
         public const string Euler = "newV[i] = prev[i] + dt * f<i>(t,prev);";
-        public static void EulerCpu(float[] prev, float[] newV, float dt, float t, int i, Func<float, float[], float> f_i)
+        public static void EulerCpu(double[] prev, double[] newV, double dt, double t, int i, Func<double, double[], double> f_i)
         {
             newV[i] = prev[i] + dt * f_i(t, prev);
         }
@@ -18,7 +18,7 @@ namespace Library
         newV[i]=tmp3+0.5f*dt*(tmp1+f<i>(t+dt,prev));
         prev[i]=tmp3;
         ";
-        public static void ImprovedEulerCpu(float[] prev, float[] newV, float dt, float t, int i, Func<float, float[], float> f_i)
+        public static void ImprovedEulerCpu(double[] prev, double[] newV, double dt, double t, int i, Func<double, double[], double> f_i)
         {
             var originalPrev = prev[i];
             var deriv = f_i(t, prev);
@@ -40,7 +40,7 @@ namespace Library
         prev[i]=tmp6;
         newV[i] = tmp6+dt/6*(tmp2+2*tmp3+2*tmp4+tmp5);
         ";
-        public static void RungeKuttaCpu(float[] prev, float[] newV, float dt, float t, int i, Func<float, float[], float> f_i)
+        public static void RungeKuttaCpu(double[] prev, double[] newV, double dt, double t, int i, Func<double, double[], double> f_i)
         {
             var dtHalf = dt / 2;
             var originalPrev = prev[i];
@@ -64,7 +64,7 @@ namespace Library
         /// <param name="by">Dimension to take derivative of</param>
         /// <param name="grid"></param>
         /// <returns>derivative approximation</returns>
-        public float Derivative(int i, int j, int k, int by, float h, float[,,] grid)
+        public double Derivative(int i, int j, int k, int by, double h, double[,,] grid)
         {
             switch (by)
             {
@@ -87,7 +87,7 @@ namespace Library
         /// <param name="k">Index 2</param>
         /// <param name="grid"></param>
         /// <returns>derivative approximation</returns>
-        public static float DerivativeZ(int i, int j, int k, float h, float[,,] grid)
+        public static double DerivativeZ(int i, int j, int k, double h, double[,,] grid)
         {
             int size2 = grid.GetLength(2);
             if (k > 1 && k < size2 - 2)
@@ -128,7 +128,7 @@ namespace Library
         /// <param name="k">Index 2</param>
         /// <param name="grid"></param>
         /// <returns>derivative approximation</returns>
-        public static float DerivativeY(int i, int j, int k, float h, float[,,] grid)
+        public static double DerivativeY(int i, int j, int k, double h, double[,,] grid)
         {
             int size1 = grid.GetLength(1);
             if (j > 1 && j < size1 - 2)
@@ -169,7 +169,7 @@ namespace Library
         /// <param name="k">Index 2</param>
         /// <param name="grid"></param>
         /// <returns>derivative approximation</returns>
-        public static float DerivativeX(int i, int j, int k, float h, float[,,] grid)
+        public static double DerivativeX(int i, int j, int k, double h, double[,,] grid)
         {
             var size0 = grid.GetLength(0);
             if (i > 1 && i < size0 - 2)
@@ -201,19 +201,19 @@ namespace Library
             );
         }
 
-        static float BackwardDifference3Points(float h, float u, float u_m, float u_mm)
+        static double BackwardDifference3Points(double h, double u, double u_m, double u_mm)
         {
             return (3 * u - 4 * u_m + u_mm) * 0.5f / h;
         }
-        static float ForwardDifference3Points(float h, float u, float u_p, float u_pp)
+        static double ForwardDifference3Points(double h, double u, double u_p, double u_pp)
         {
             return (-3 * u + 4 * u_p - u_pp) * 0.5f / h;
         }
-        static float CentralDifference(float h, float u_p, float u_m)
+        static double CentralDifference(double h, double u_p, double u_m)
         {
             return (u_p - u_m) * 0.5f / h;
         }
-        static float CentralDifference4Order(float h, float u_p, float u_pp, float u_m, float u_mm)
+        static double CentralDifference4Order(double h, double u_p, double u_pp, double u_m, double u_mm)
         {
             return 0.0833333333f / h * (8 * (u_p - u_m) + u_mm - u_pp);
         }
